判断设备是否在线 ;计算已花费租金

+++ provider/src/main/java/com/ycb/wxxcx/provider/controller/ShopController.java
@@ -55,7 +55,7 @@ public class ShopController {
     @ResponseBody
     // @Action --修改 为拦截器方式实现
     public String query(@RequestParam("session") String session,
-                        @RequestParam("shop_id") Integer shopid) {
+                        @RequestParam("shop_id") Long shopid) {
         Map<String, Object> bacMap = new HashMap<>();
         try {
             ShopStation shop =  this.shopMapper.findShopInfo(shopid);
